+++ CHANGELOG.md
@@ -6,9 +6,10 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 0.8.11
 
-2020 Apr 4 [in progress]
+2020 Apr 6 [in progress]
 
 *   Fixed waveform Y-Axis zoom only zooms out with trackpad #70.
+*   Fixed PySide2 imports #71.
 
 
 ## 0.8.10